Shuffle code in header file to new header file. Updates #includes.
