+++ flang/lib/Optimizer/Transforms/ControlFlowConverter.cpp
@@ -21,11 +21,13 @@
 #include "flang/Optimizer/Dialect/FIRType.h"
 #include "flang/Optimizer/Transforms/Passes.h"
 #include "mlir/Conversion/AffineToStandard/AffineToStandard.h"
+#include "mlir/Dialect/SCF/SCF.h"
 #include "mlir/Dialect/StandardOps/IR/Ops.h"
 #include "mlir/IR/BuiltinTypes.h"
 #include "mlir/Pass/Pass.h"
 #include "mlir/Transforms/DialectConversion.h"
-#include "llvm/ADT/ArrayRef.h"
+
+#define DEBUG_TYPE "flang-cfg-converter"
 
 static llvm::cl::opt<bool> disableControlFlowLowering(
     "disable-control-flow-lowering",
